Minor: Compile + Minify Scripts/Styles
This is actually a pretty major update, but it's a bit too technical and
meta to go in my site's RSS feed.

This change is a long time coming: now we're using TypeScript for major
site scripts and Sass for major site styles. We use Gulp to
compile+minify our sources. The output files are configured to live next
to the input files, so "src/script/nav.ts" compiles to
"src/script/nav.js". Maybe one day if we can get it working, we'll also
use imagemin to compress images too.

Also I've removed the old "markdown.min.js" and replaced it with an npm
dependency: snarkdown. It is much smaller, and is installed as a
dependency so its minified source doesn't have to live in source
control!

oh also i added some more links and stuff idk