Minor: Fix scripts that rely on external content
These scripts throw errors when the user is not connected to the
internet. On the actual website that's fine and expected behavior, but
in theory the end to end tests should be able to pass without an
internet connection.

Ideally there should be some errors shown on the page if these errors
are caught, but for now they fail silently. ... I'll work on it
eventually.